Fix Darwin build: use `size_t` for `boost::hash_combine` seed
On Apple platforms, `std::size_t` is `unsigned long` while `UInt64` is
`unsigned long long` — different types of the same width — so the
non-const reference parameter of `boost::hash_combine` does not bind to
a `UInt64` lvalue and the build fails with:

    error: no matching function for call to 'hash_combine'
    note: candidate function template not viable: no known conversion
          from 'UInt64' (aka 'unsigned long long') to 'std::size_t &'
          (aka 'unsigned long &') for 1st argument

On x86_64 Linux both aliases happen to resolve to `unsigned long`, which
is why CI Linux builds passed but `Build (amd_darwin)` and
`Build (arm_darwin)` failed.

Switch the local seed variable on both the read path
(`MergeTreeDataSelectExecutor::filterPartsByQueryConditionCache`) and
the write path (`updateQueryConditionCache`) to `size_t`. The implicit
conversions to/from `UInt64` at the cache-API boundary are lossless on
all 64-bit platforms ClickHouse supports.

CI report: https://s3.amazonaws.com/clickhouse-test-reports/json.html?PR=104478&sha=4ab18afb9c043a670ca72211ab85494fd956e3b8&name_0=PR&name_1=Build%20%28amd_darwin%29
PR: https://github.com/ClickHouse/ClickHouse/pull/104478

+++ src/Processors/QueryPlan/Optimizations/updateQueryConditionCache.cpp
@@ -57,7 +57,9 @@ void updateQueryConditionCache(const Stack & stack, const QueryPlanOptimizationS
     {
         if (auto * filter_step = typeid_cast<FilterStep *>(iter->node->step.get()))
         {
-            UInt64 condition_hash = filter_actions_dag->getOutputs()[0]->getHash();
+            /// `size_t` (not `UInt64`) so `boost::hash_combine` binds on platforms where
+            /// they differ (e.g. Apple, where `size_t` is `unsigned long` but `UInt64` is `unsigned long long`).
+            size_t condition_hash = filter_actions_dag->getOutputs()[0]->getHash();
 
             /// `ORDER BY ... LIMIT N` may drop granules during reading, so the result of the WHERE
             /// filter is no longer "applies to every granule of every part" — it applies only to
